+++ src/model/shape/Rectangle.java
@@ -10,7 +10,7 @@ import model.utils.Time;
  */
 public class Rectangle extends AShape {
 
-  public Rectangle(String name, Shape type, Color col, int posX, int posY,
+  public Rectangle(String name, Shape type, Color col, double posX, double posY,
                    int width, int height, Time time) {
     super(name, type, col, posX, posY, width, height, time);
   }
